#14 remove all logic from goferbot.js
