+++ goferbot.js
@@ -5,21 +5,14 @@ var Wit = require('node-wit').Wit;
 var dbmanager = require('./lib/dbmanager.js');
 var logic = require('./lib/witlogic.js');
 var takeOrders = require('./lib/takeOrder.js');
+var openConvo = require('./lib/openConvo.js');
+var orderTicket = require('./lib/orderTicket.js')
 
 var wit = new Wit(process.env.WIT_TOKEN, logic.actions);
 
 module.exports = function(controller, botkit){
-  controller.on('rtm_open', takeOrders); // roll call for everyone opting in for coffee run
-  controller.hears('pick up', 'direct_message', getOrderList); // respond to pick up wants
-  controller.on('direct_message', require('./lib/openConvo.js'));
-}
+  controller.on('rtm_open', takeOrders);
+  controller.on('direct_message', openConvo);
 
-function getOrderList(bot, message) {
- var list = dbmanager.get(message.team);
- list.then(function(tab) {
-   var text = tab.map(function(order) {
-     return (order.real_name || order.name || order.user) + ' wants `' + order.order + '`';
-   }).join('\n')
-   bot.reply(message, text);
- });
+  controller.hears('pick up', 'direct_message', orderTicket);
 }
